Implementation of soldier class restrictions.
An option has been added to give the user control over the soldier
classes that are included in the overview stats plots.

A function to create a copy of an overview dictionary was added to allow
for the implementation of the above feature.

Additionally the directories functions have been renamed and configured
to support the creation of directories outside the pwd of the call to
python. There was a bug that prevented the creation of directories
outside the pwd as a complete directory reference for example, /home/...
meant that when splitting the name at each '/' char resulted in the
first directory created being ''. This was fixed by including an if
statement to change a '' on the split directory name to be '/'.

+++ battlefield/directories.py
@@ -1,6 +1,6 @@
 # Author: Jay Oliver
 # Date Created: 03/05/2020
-# Date Last Modified: 03/05/2020
+# Date Last Modified: 05/05/2020
 # Purpose: Contains functions that pertain to directory manipulation
 # Comments:
 
@@ -8,7 +8,7 @@ from os import mkdir
 from os import rmdir
 
 
-def create_pwd(direct):
+def create_dir(direct):
     """Create the given directory in the present working directory.
 
     The present working directory is taken as the location of invokation.
@@ -19,7 +19,7 @@ def create_pwd(direct):
                  the structure is as follows:
                     |path  (direct parameter eg. "some/thing")
                     |0     (The first file in the path. A 0 is stored if the
-                    |1        creation of the file failedl, a 1 on success.)
+                    |1        creation of the file failed, a 1 on success.)
                     ...
     raises:
         - TypeError: When the given parameter is not of the expected type
@@ -34,6 +34,8 @@ def create_pwd(direct):
     path = ""
     for ind, val in enumerate(direct.split('/')):
         try:
+            if val == '':
+                val = '/'
             path = path + val
             mkdir(path)
             files[ind] = 1
@@ -43,7 +45,7 @@ def create_pwd(direct):
     return files
 
 
-def remove_pwd(files):
+def remove_dir(files):
     """Remove the files marked as created by create_pwd
 
     parameters:
@@ -51,7 +53,7 @@ def remove_pwd(files):
                  created. The structure is as follows:
                     |path  (direct parameter eg. "some/thing")
                     |0     (The first file in the path. A 0 is stored if the
-                    |1        creation of the file failedl, a 1 on success.)
+                    |1        creation of the file failed, a 1 on success.)
                     ...
     raises:
         - TypeError: When the given parameter does not meet the expected type
@@ -71,7 +73,12 @@ def remove_pwd(files):
         if files[i] == 1:
             try:
                 rm_file = "/".join(files['path'].split("/")[:i + 1:])
-                rmdir(rm_file)
+                if rm_file == "/":
+                    pass
+                elif rm_file == "/home":
+                    pass
+                else:
+                    rmdir(rm_file)
             except OSError:
                 raise OSError("A directory that was marked as created is not"
                               " empty, {}".format(rm_file))
